Rename and clean up logging and starlette
