+++ pyproject.toml
@@ -7,7 +7,6 @@ readme = "README.md"
 
 [tool.poetry.dependencies]
 python = "^3.11"
-jsonformatter = "^0.3.2"
 opentelemetry-api = { version = "^1.28.1", optional = true }
 opentelemetry-exporter-otlp = { version = "^1.28.1", optional = true }
 opentelemetry-instrumentation-botocore = { version = ">=0.48b0", optional = true }
